# Conflicts:
#	drivers/modem/modem_cellular.c
#	drivers/wifi/eswifi/eswifi_socket_offload.c
#	include/zephyr/drivers/cellular.h
#	include/zephyr/net/net_context.h
#	subsys/modem/modem_cmux.c
#	subsys/modem/modem_ppp.c
#	subsys/net/ip/tcp.c
#	subsys/net/l2/wifi/wifi_mgmt.c
#	subsys/net/lib/mqtt_sn/mqtt_sn_transport_udp.c
#	subsys/net/lib/sockets/sockets_inet.c
#	subsys/tracing/ctf/ctf_top.c
#	subsys/tracing/ctf/tracing_ctf.h

+++ include/zephyr/net/net_context.h
@@ -1398,7 +1398,7 @@ enum net_context_option {
  */
 int net_context_set_option(struct net_context* context,
                            enum net_context_option option,
-                           void const* value, size_t len);
+                           void const* value, uint32_t len);
 
 /**
  * @brief Get connection option value for this context.
@@ -1412,7 +1412,7 @@ int net_context_set_option(struct net_context* context,
  */
 int net_context_get_option(struct net_context* context,
                            enum net_context_option option,
-                           void* value, size_t* len);
+                           void* value, uint32_t* len);
 
 /**
  * @typedef net_context_cb_t
